Tiny fix in strlen

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: hmorand <hmorand@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/11 13:45:00 by hmorand           #+#    #+#             */
-/*   Updated: 2023/08/14 10:41:09 by hmorand          ###   ########.fr       */
+/*   Updated: 2023/08/14 16:13:30 by hmorand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ size_t	ft_strlcat(char *dest, const char *src, size_t n)
 	}
 	if (i < n)
 		dest[i + j] = '\0';
-	return (i + (size_t)ft_strlen(src));
+	return (i + ft_strlen(src));
 }
 
 /* #include <stdio.h>
