+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: hmorand <hmorand@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/09 11:19:18 by hmorand           #+#    #+#             */
-/*   Updated: 2023/08/14 10:41:21 by hmorand          ###   ########.fr       */
+/*   Updated: 2023/08/14 16:11:29 by hmorand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
